Super simple scripts
Don't run the scripts in place anymore.  Rely on setting environment
variables to set parameters for dbgen and the postgres scripts.  The
other DBMS systems' scripts will also need to be updated.  This will
hopefully make it easier to work with multiple databases without having
to rebuild all the scripts in order to switch.