Update report script
Use the data collected during the test instead of checking the system
that the script runs from.  The report generator may not be run on the
same system as the test was, so certain system details were of the
system running the report as opposed to the system under test.