PostgreSQL patches for dbgen
Add patches so for dbgen to support PostgreSQL.  Also include a patch
so that loading can be streamed directly into the database.

The load and power test scripts needed to be updated to use environment
variables that define where the dbgen and qgen binaries are, since there
are no Makefile rules for installing dbgen and qgen were not reapplied.