Remove -g flag from run script
It's not really necessary at this level.  For ease of use, which is what
the "run" script is designed for, rely on the scripts to only generate
files if they already don't exist.

If the user wants finer control to regenerate flat files, they can be
regenerated by running the load script manually or by deleting the DSS
directory to clean up all the generated files.