appimage: update handling of dbgen
First it wasn't working because the environment variable was named EGEN
incorrectly.

Now set DBGEN the dbgen environment variable for the name of the dbgen
zipfile, which will be encoded with your serial number if it hasn't been
renamed.  The scripts currently will expect the zip file to be in the
top level repository directly.